requirements.txt
torch>=1.12
numpy
pandas
scikit-learn
tqdm
biopython
transformers
fair-esm
